requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
telethon>=1.33.1
python-dotenv>=1.0.0
stripe>=7.10.0
httpx>=0.26.0
pydantic>=2.6.0
slowapi>=0.1.9
python-multipart>=0.0.7
email-validator>=2.1.0
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
alembic
redis
celery
